refactor: 기존 View를 InputView, OutputView로 분리 후 패키지 이동
