+++ src/main/kotlin/racingcar/Application.kt
@@ -1,9 +1,10 @@
 package racingcar
 
 import racingcar.controller.CarController
+import racingcar.view.InputView
 
 fun main() {
-    val view = View()
+    val view = InputView()
     val model = Model()
     val controller = CarController(model, view)
     controller.run()
